Workaround for hitbox updating glitches with transformed objects, removed some old debugging stuff, simplified stuff
The issue was because the transform was a reference, which could change while
painting the hitbox because painting is done in another thread and
AffineTransform updates aren't atomic. The naive way to fix this would be to
always clone the transforms while updating them, but this would cause
significant lag due to Java's slow memory allocator and the extra GC cycles
needed for the objects. The slightly better way is to clone them if debugging
is enabled, but to an existing AffineTransform object used only for drawing
hitboxes.